+++ apps/web/src/store/index.ts
@@ -1,0 +1,19 @@
+import { statusWebsiteSlice } from "./status";
+import { bridgeApi } from "./website";
+
+const {
+  useGenerateAddressMutation,
+  useLazyVerifyQuery,
+  useGetAddressDetailMutation,
+} = bridgeApi;
+
+const { useGetBridgeStatusQuery } = statusWebsiteSlice;
+
+export {
+  useGetBridgeStatusQuery,
+  useGenerateAddressMutation,
+  useLazyVerifyQuery,
+  useGetAddressDetailMutation,
+  bridgeApi,
+  statusWebsiteSlice,
+};
